==> scripts/04.5_getting_keywords_sleep_related.ipynb <==
import pandas as pd
from collections import Counter
from nltk.util import ngrams
from nltk.tokenize import word_tokenize
import nltk
from nltk.corpus import stopwords
import string
nltk.download('punkt_tab')

# Download tokenizer models if not already available
nltk.download('punkt')


def flatten_list_of_lists(nested_list):
    return [item for sublist in nested_list for item in sublist]


# Ensure required resources are downloaded
nltk.download('punkt')
nltk.download('stopwords')
df = pd.read_pickle("../exports/04b_groundtruth/P-SET/Annotated/Sleep_Sentences.pkl")
def most_frequent_ngrams(sentences, N=5):
    stop_words = set(stopwords.words('english'))
    punctuation = set(string.punctuation)

    all_unigrams = []
    all_bigrams = []
    all_trigrams = []

    for sentence in sentences:
        tokens = word_tokenize(sentence.lower())
        # Remove stopwords and punctuation
        filtered_tokens = [t for t in tokens if t not in stop_words and t not in punctuation]

        all_unigrams.extend(filtered_tokens)
        all_bigrams.extend(ngrams(filtered_tokens, 2))
        all_trigrams.extend(ngrams(filtered_tokens, 3))

    # Count frequencies
    unigram_counts = Counter(all_unigrams)
    bigram_counts = Counter(all_bigrams)
    trigram_counts = Counter(all_trigrams)

    # Get top N most common
    most_common_unigram = unigram_counts.most_common(N)
    most_common_bigram = bigram_counts.most_common(N)
    most_common_trigram = trigram_counts.most_common(N)

    return {
        'most_common_unigram': most_common_unigram,
        'most_common_bigram': most_common_bigram,
        'most_common_trigram': most_common_trigram
    }
most_frequent_ngrams(df[df.is_keyword_present==True]['Sentence'],30)

drug_names = [ 

    "zolpidem", "Ambien", "temazepam", "Lunesta", "Ativan", "lorazepam", "Restoril", 

    "flurazepam", "estazolam", "doxepin", "Halcion", "Belsomra", "diphenhydramine", 

    "eszopiclone", "quazepam", "Ambien CR", "Rozerem", "Silenor", "Sonata", 

    "doxylamine", "Edluar", "triazolam", "Compoz Nighttime Sleep Aid", "Doral", 

    "zopiclone", "ramelteon", "Unisom SleepTabs", "zaleplon", "Dayvigo", "Quviviq", 

    "suvorexant", "Unisom SleepMelts", "Zolpimist", "Nighttime Sleepaid", "Nytol", 

    "Simply Sleep", "Sleep Aid", "Sominex", "Unisom SleepGels", "ZzzQuil", 

    "daridorexant", "Dormin", "lemborexant", "Nytol QuickCaps", "Sleep-ettes D", 

    "Sleepinal", "Twilite", "Z-Sleep" 

] 

len(drug_names)

df[['Sent_gt_Sleep','is_keyword_present']].value_counts()
